agents/requirements.txt
# LangChain and LangSmith - use older stable versions for Python 3.12
langchain==0.1.20
langchain-openai==0.1.8
langsmith==0.1.65

# API and HTTP clients
requests==2.31.0
httpx==0.26.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data processing
python-dotenv==1.0.0
pydantic==2.5.3

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
